Splitting tests and updating for new features
