+++ test/model.js
@@ -14,16 +14,16 @@ const types = {
   string: 'varchar',
   number: 'integer',
   boolean: 'boolean',
-  datetime: 'timestamp with time zone',
-  text: 'text',
-  json: 'jsonb',
+  datetime: { js: 'string', pg: 'timestamp with time zone' },
+  text: { js: 'string', pg: 'text' },
+  json: { js: 'schema', pg: 'jsonb' },
 };
 
 metatests.test('Model: from struct', (test) => {
   const entities = new Map();
 
   entities.set('Company', {
-    Dictionary: {},
+    Dictionary: { store: 'persistent', scope: 'application' },
     name: { type: 'string', unique: true },
     addresses: { many: 'Address' },
   });
@@ -37,14 +37,15 @@ metatests.test('Model: from struct', (test) => {
     driver: 'pg',
   });
 
-  test.strictEqual(model.types, {
-    boolean: 'boolean',
-    datetime: 'timestamp with time zone',
-    json: 'jsonb',
-    number: 'integer',
-    string: 'varchar',
-    text: 'text',
-  });
+  const { string, number, boolean } = model.types;
+  test.strictEqual(string.prototype.pg, types.string);
+  test.strictEqual(number.prototype.pg, types.number);
+  test.strictEqual(boolean.prototype.pg, types.boolean);
+
+  const { datetime, text, json } = model.types;
+  test.strictEqual(datetime.prototype.pg, types.datetime.pg);
+  test.strictEqual(text.prototype.pg, types.text.pg);
+  test.strictEqual(json.prototype.pg, types.json.pg);
 
   test.strictEqual(model.order, new Set(['Company']));
 
@@ -55,9 +56,10 @@ metatests.test('Model: from struct', (test) => {
   test.strictEqual(company.store, 'persistent');
   test.strictEqual(company.scope, 'application');
 
-  test.strictEqual(company.fields, {
-    name: { type: 'string', unique: true, required: true },
-  });
+  const { name } = company.fields;
+  test.strictEqual(name.type, 'string');
+  test.strictEqual(name.required, true);
+  test.strictEqual(name.unique, true);
 
   const warn = model.warnings[0];
   test.strictEqual(
@@ -65,26 +67,5 @@ metatests.test('Model: from struct', (test) => {
     'Warning: "Address" referenced by "Company" is not found'
   );
 
-  test.end();
-});
-
-metatests.test('Model: loader', async (test) => {
-  const model = await Model.load(process.cwd() + '/test/schemas', types);
-  test.strictEqual(model.entities.size, 5);
-  const Account = model.entities.get('Account');
-  test.strictEqual(Account.fields.fullName.type, 'schema');
-  test.strictEqual(Account.fields.fullName.schema.constructor.name, 'Schema');
-  test.strictEqual(model.order.size, 5);
-  test.strictEqual(typeof model.types, 'object');
-  test.strictEqual(typeof model.database, 'object');
-  test.end();
-});
-
-metatests.test(`Model: restricted 'type' property`, (test) => {
-  const model = new Model(
-    { string: 'string' },
-    new Map([['FailingEntity', { type: 'string' }]])
-  );
-  test.strictEqual(model.warnings.length, 1);
   test.end();
 });
